+++ AutoSaver/AutoSaver.cs
@@ -1,14 +1,17 @@
 ﻿using System.Globalization;
 using BepInEx;
 using BepInEx.Bootstrap;
+using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
 using Marioalexsan.AutoSaver.HarmonyReversePatches;
+using Marioalexsan.AutoSaver.SoftDependencies;
 using UnityEngine;
 
 namespace Marioalexsan.AutoSaver;
 
 [BepInPlugin(ModInfo.PLUGIN_GUID, ModInfo.PLUGIN_NAME, ModInfo.PLUGIN_VERSION)]
+[BepInDependency(EasySettings.ModID, BepInDependency.DependencyFlags.SoftDependency)]
 public class AutoSaver : BaseUnityPlugin
 {
     public const string MoreBankTabsIndentifier = "com.16mb.morebanktabs";
@@ -18,121 +21,224 @@ public class AutoSaver : BaseUnityPlugin
 
     public new ManualLogSource Logger { get; private set; }
 
-    private Harmony _harmony;
-    private TimeSpan _elapsedTime;
+    private readonly Harmony _harmony;
+    private TimeSpan _elapsedAutosaveTime;
 
-    public bool SaveOnMapChange { get; private set; }
-    public bool CharacterActive { get; private set; }
-    public bool AutosaveActive => CharacterActive;
-    public bool EnableExperimentalFeatures { get; private set; }
-    public KeyCode SaveMultiplayerKeyCode { get; private set; }
+    public ConfigEntry<bool> EnableAutosaving { get; private set; }
+    public ConfigEntry<bool> SaveOnMapChange { get; private set; }
+    public ConfigEntry<bool> EnableExperimentalFeatures { get; private set; }
+    public ConfigEntry<bool> AppendSlotToSaveName { get; private set; }
+    public ConfigEntry<KeyCode> SaveMultiplayerKeyCode { get; private set; }
+
+    public ConfigEntry<int> BackupIntervalInMinutesConfig { get; private set; }
+    public ConfigEntry<int> SaveCountToKeepConfig { get; private set; }
+
+    public int BackupIntervalInMinutes { get; private set; }
+    public int SaveCountToKeep { get; private set; }
+
+    public bool IsInGame { get; private set; }
+
+    public TimeSpan AutosaveInterval => TimeSpan.FromMinutes(BackupIntervalInMinutes);
+
+    public bool DetectedMoreBankTabsMod { get; private set; } = false;
 
     private readonly char[] BannedChars = [.. Path.GetInvalidPathChars(), .. Path.GetInvalidFileNameChars()];
+
+    private string ObsoleteModDataFolderPath => Path.Combine(ProfileDataManager._current._dataPath, ModDataFolderName);
+    private readonly string ModDataFolderName = "Marioalexsan_AutoSaver";
+    private string ModDataFolderPath => Path.Combine(Path.GetDirectoryName(Paths.ExecutablePath), ModDataFolderName);
+    private string CharacterFolderPath => Path.Combine(ModDataFolderPath, "Characters");
+    private string ItemBankFolderPath => Path.Combine(ModDataFolderPath, "ItemBank");
+    private string MultiSavesFolderPath => Path.Combine(ModDataFolderPath, "Multi");
+
+    private bool _checkedForOldAutosaverBackups = false;
 
     public AutoSaver()
     {
         _plugin = this;
         Logger = base.Logger;
         _harmony = new Harmony(ModInfo.PLUGIN_GUID);
+
+        EnableAutosaving = Config.Bind("General", "EnableAutosaving", true, "True to enable autosaving every few minutes, false to disable it.");
+        SaveOnMapChange = Config.Bind("General", "SaveOnMapChange", false, "Set to true to trigger autosaving whenever a new level is loaded.");
+        AppendSlotToSaveName = Config.Bind("General", "AppendSlotToSaveName", true, "Set to true to append slot index to the saved character names, false to use the character name only. Recommended if you use characters with duplicate names.");
+        BackupIntervalInMinutesConfig = Config.Bind("General", "BackupInterval", 4, new ConfigDescription("Interval between save backups, in minutes.", new AcceptableValueRange<int>(1, 60))); // Has to be float range due to EasySettings
+        SaveCountToKeepConfig = Config.Bind("General", "SavesToKeep", 15, new ConfigDescription("Maximum number of saves to keep.", new AcceptableValueRange<int>(5, 50))); // Has to be float range due to EasySettings
+
+        BackupIntervalInMinutes = BackupIntervalInMinutesConfig.Value;
+        SaveCountToKeep = SaveCountToKeepConfig.Value;
+
+        EnableExperimentalFeatures = Config.Bind("Experimental", "EnableExperimentalFeatures", false, "Set to true to enable experimental features.");
+        SaveMultiplayerKeyCode = Config.Bind(
+            "Experimental", 
+            "SaveMultiplayerKeyCode", 
+            KeyCode.KeypadPlus,
+            $"Key to trigger saving other people's saves in multiplayer under the \"Multi\" folder.{Environment.NewLine}Note that saves saved in this way are lackluster."
+        );
     }
 
     private void Awake()
     {
-        Logger.LogInfo("AutoSaver patching...");
         _harmony.PatchAll();
+        InitializeConfiguration();
 
-        Logger.LogInfo("AutoSaver configuring...");
-        Configure();
-
-        Logger.LogInfo("AutoSaver initialized!");
+        Logging.LogInfo("AutoSaver initialized!");
     }
 
-    private TimeSpan _autosaveInterval;
-    private int _saveCountToKeep;
-
-    public bool DetectedMoreBankTabsMod { get; private set; } = false;
-
-    private void Configure()
+    private void CheckForOldAutosaverBackups()
     {
-        int autosaveMinutes = Config.Bind("General", "BackupInterval", 4, "Interval between save backups, in minutes (min 1, max 60).").Value;
+        if (Directory.Exists(ObsoleteModDataFolderPath))
+        {
+            if (Directory.Exists(ModDataFolderPath))
+            {
+                Logging.LogWarning($"Found old Autosaver folder in {ObsoleteModDataFolderPath}, but new save location {ModDataFolderPath} already exists.");
+                Logging.LogWarning($"Please backup and move {ObsoleteModDataFolderPath} out of the profileCollections folder manually!");
+                return;
+            }
 
-        autosaveMinutes = Math.Max(Math.Min(autosaveMinutes, 60), 1);
-        _autosaveInterval = TimeSpan.FromMinutes(autosaveMinutes);
+            Logging.LogInfo($"Found old Autosaver folder in {ObsoleteModDataFolderPath}, will try to move it to the new save location {ModDataFolderPath}.");
 
-        _saveCountToKeep = Config.Bind("General", "SavesToKeep", 15, "Maximum number of saves to keep (min 5, max 50).").Value;
-        _saveCountToKeep = Math.Max(Math.Min(_saveCountToKeep, 50), 5);
+            try
+            {
+                Directory.Move(ObsoleteModDataFolderPath, ModDataFolderPath);
+                Logging.LogInfo($"Backup data moved successfully to the new location!");
+            }
+            catch (Exception e)
+            {
+                Logging.LogError($"Failed to move old backup folder! Please send the exception message to the mod developer:");
+                Logging.LogError(e.ToString());
+                Logging.LogError($"Please check {ObsoleteModDataFolderPath} and backup and move it out of the profileCollections folder manually!");
+            }
+        }
+    }
 
-        SaveOnMapChange = Config.Bind("General", "SaveOnMapChange", false, "Set to true to trigger autosaving whenever a new level is loaded.").Value;
+    private void InitializeConfiguration()
+    {
+        if (EasySettings.IsAvailable)
+        {
+            EasySettings.OnApplySettings.AddListener(() =>
+            {
+                try
+                {
+                    Config.Save();
 
-        EnableExperimentalFeatures = Config.Bind("Experimental", "EnableExperimentalFeatures", false, "Set to true to enable experimental features.").Value;
-        SaveMultiplayerKeyCode = Config.Bind("Experimental", "SaveMultiplayerKeyCode", KeyCode.KeypadPlus, $"Key to trigger saving other people's saves in multiplayer under the \"Multi\" folder.{Environment.NewLine}Note that saves saved in this way are lackluster.").Value;
+                    // This is set on apply since otherwise it would cause things
+                    // to get garbage collected or saved ahead of time if we take the values it applies on the fly
+                    BackupIntervalInMinutes = BackupIntervalInMinutesConfig.Value;
+                    SaveCountToKeep = SaveCountToKeepConfig.Value;
+                }
+                catch (Exception e)
+                {
+                    Logging.LogError($"AutoSaevr crashed in OnApplySettings! Please report this error to the mod developer:");
+                    Logging.LogError(e.ToString());
+                }
+            });
+            EasySettings.OnInitialized.AddListener(() =>
+            {
+                EasySettings.AddHeader(ModInfo.PLUGIN_NAME);
+                EasySettings.AddToggle("Enable periodic autosaves", EnableAutosaving);
+                EasySettings.AddToggle("Enable autosaving whenever map changes", SaveOnMapChange);
+                EasySettings.AddToggle("Append save slot to save names (recommended)", AppendSlotToSaveName);
+                // EnableExperimentalFeatures is skipped on purpose
+                EasySettings.AddAdvancedSlider("Backup interval (minutes)", BackupIntervalInMinutesConfig);
+                EasySettings.AddAdvancedSlider("Number of saves to keep", SaveCountToKeepConfig);
+                // SaveMultiplayerKeyCode is skipped on purpose
+
+                EasySettings.AddButton("Open save backups folder", () =>
+                {
+                    Application.OpenURL(new Uri($"{ModDataFolderPath}").AbsoluteUri);
+                });
+            });
+        }
     }
 
     internal void GameEntered()
     {
-        Logger.LogInfo("Game entered / map switched. Activating character autosaves.");
-        ProfileDataManager._current.Load_ItemStorageData(); // This isn't loaded on start for some reason
-        RunAutosaves();
-        _elapsedTime = TimeSpan.Zero;
-        CharacterActive = true;
+        bool runSave = false;
+
+        if (!IsInGame)
+        {
+            IsInGame = true;
+            runSave = true;
+            Logging.LogInfo("Game entered. Activating character autosaves.");
+        }
+        else if (SaveOnMapChange.Value)
+        {
+            runSave = true;
+            Logging.LogInfo("Map changed. Autosaving.");
+        }
+
+        if (runSave)
+        {
+            ProfileDataManager._current.Load_ItemStorageData(); // This isn't loaded on start for some reason
+            RunAutosaves();
+        }
     }
 
     internal void GameExited()
     {
-        Logger.LogInfo("Game exited. Stopping character autosaves.");
+        Logging.LogInfo("Game exited. Stopping character autosaves.");
         RunAutosaves();
-        _elapsedTime = TimeSpan.Zero;
-        CharacterActive = false;
+        IsInGame = false;
     }
 
     private void Update()
     {
+        // Cannot be done in Awake since it relies on dataPath to be available
+        if (!_checkedForOldAutosaverBackups && ProfileDataManager._current?._dataPath != null)
+        {
+            _checkedForOldAutosaverBackups = true;
+            CheckForOldAutosaverBackups();
+        }
+
         if (!DetectedMoreBankTabsMod)
         {
             if (Chainloader.PluginInfos.ContainsKey(MoreBankTabsIndentifier))
             {
                 DetectedMoreBankTabsMod = true;
-                Logger.LogInfo($"Detected MoreBankTabs mod ({MoreBankTabsIndentifier}).");
-                Logger.LogInfo($"Will try to backup the extra bank tabs.");
+                Logging.LogInfo($"Detected MoreBankTabs mod ({MoreBankTabsIndentifier}).");
+                Logging.LogInfo($"Will try to backup the extra bank tabs.");
             }
         }
 
-        if (AutosaveActive)
+        if (IsInGame)
         {
-            _elapsedTime += TimeSpan.FromSeconds(Time.deltaTime);
+            _elapsedAutosaveTime += TimeSpan.FromSeconds(Time.deltaTime);
 
-            if (_elapsedTime >= _autosaveInterval)
+            if (EnableAutosaving.Value)
             {
-                _elapsedTime = TimeSpan.Zero;
-                RunAutosaves();
+                if (_elapsedAutosaveTime >= AutosaveInterval)
+                {
+                    Logging.LogInfo($"Periodic autosave triggered ({AutosaveInterval.TotalMinutes} minutes have passed).");
+                    RunAutosaves();
+                }
             }
-            else
+
+            // Retry if player is buffering or w/e, and the last round of saving didn't finish
+            // This applies to autosaves, and map changes
+            if (!CharacterAutoSaver.SaveDone)
             {
-                // Retry if player is buffering or w/e
-                if (!CharacterAutoSaver.SaveDone)
-                {
-                    AutosaveCurrentCharacter();
-                }
-                if (!ItemBankAutoSaver.SaveDone)
-                {
-                    AutosaveCurrentItemBank();
-                }
+                AutosaveCurrentCharacter();
+            }
+            if (!ItemBankAutoSaver.SaveDone)
+            {
+                AutosaveCurrentItemBank();
             }
         }
 
-        if (EnableExperimentalFeatures && Input.GetKeyDown(SaveMultiplayerKeyCode))
+        if (EnableExperimentalFeatures.Value && Input.GetKeyDown(SaveMultiplayerKeyCode.Value))
         {
-            Logger.LogInfo("[Experimental] Saving every online player's saves.");
+            Logging.LogInfo("[Experimental] Saving every online player's saves.");
             
             foreach (var player in FindObjectsOfType<Player>())
             {
-                Logger.LogInfo($"Saving player data for {player._nickname}");
+                Logging.LogInfo($"Saving player data for {player._nickname}");
                 Directory.CreateDirectory(MultiSavesFolderPath);
                 CharacterAutoSaver.TrySaveSpecificProfileToLocation(player, Path.Combine(MultiSavesFolderPath, SanitizePlayerName(player)));
 
                 if (!CharacterAutoSaver.SaveDone)
                 {
-                    Logger.LogWarning($"Failed to do save for {player._nickname}. Current game status is {player._currentGameCondition}.");
+                    Logging.LogWarning($"Failed to do save for {player._nickname}. Current game status is {player._currentGameCondition}.");
                 }
             }
         }
@@ -140,6 +246,7 @@ public class AutoSaver : BaseUnityPlugin
 
     private void RunAutosaves()
     {
+        _elapsedAutosaveTime = TimeSpan.Zero;
         AutosaveCurrentCharacter();
         AutosaveCurrentItemBank();
     }
@@ -174,20 +281,21 @@ public class AutoSaver : BaseUnityPlugin
 
     public string GetBackupNameForCurrentPlayer()
     {
-        return $"{SanitizePlayerName(Player._mainPlayer)}_slot{ProfileDataManager._current._selectedFileIndex}";
+        if (AppendSlotToSaveName.Value)
+        {
+            return $"{SanitizePlayerName(Player._mainPlayer)}_slot{ProfileDataManager._current._selectedFileIndex}";
+        }
+        else
+        {
+            return $"{SanitizePlayerName(Player._mainPlayer)}";
+        }
     }
-
-    private readonly string ModDataFolderName = "Marioalexsan_AutoSaver";
-    private string ModDataFolderPath => Path.Combine(ProfileDataManager._current._dataPath, ModDataFolderName);
-    private string CharacterFolderPath => Path.Combine(ModDataFolderPath, "Characters");
-    private string ItemBankFolderPath => Path.Combine(ModDataFolderPath, "ItemBank");
-    private string MultiSavesFolderPath => Path.Combine(ModDataFolderPath, "Multi");
 
     private void RunItemBankGarbageCollector()
     {
         if (!Directory.Exists(ItemBankFolderPath))
         {
-            Logger.LogWarning($"Couldn't find backup path {ItemBankFolderPath} to run garbage collection for.");
+            Logging.LogWarning($"Couldn't find backup path {ItemBankFolderPath} to run garbage collection for.");
             return;
         }
 
@@ -200,7 +308,7 @@ public class AutoSaver : BaseUnityPlugin
         names.Remove("_latest");
         names.Sort();
 
-        while (names.Count > _saveCountToKeep)
+        while (names.Count > SaveCountToKeep)
         {
             var saveToDelete = names[0];
             names.RemoveAt(0);
@@ -221,7 +329,7 @@ public class AutoSaver : BaseUnityPlugin
 
         if (!Directory.Exists(backupPath))
         {
-            Logger.LogWarning($"Couldn't find backup path {backupPath} to run garbage collection for.");
+            Logging.LogWarning($"Couldn't find backup path {backupPath} to run garbage collection for.");
             return;
         }    
 
@@ -234,7 +342,7 @@ public class AutoSaver : BaseUnityPlugin
         names.Remove("_latest");
         names.Sort();
 
-        while (names.Count > _saveCountToKeep)
+        while (names.Count > SaveCountToKeep)
         {
             var saveToDelete = names[0];
             names.RemoveAt(0);
@@ -277,8 +385,8 @@ public class AutoSaver : BaseUnityPlugin
         }
         catch (Exception e)
         {
-            Logger.LogError($"Failed to autosave!");
-            Logger.LogError($"Exception message: {e}");
+            Logging.LogError($"Failed to autosave!");
+            Logging.LogError($"Exception message: {e}");
             return;
         }
     }
@@ -289,7 +397,7 @@ public class AutoSaver : BaseUnityPlugin
         {
             if (!(bool)Player._mainPlayer)
             {
-                Logger.LogError($"Couldn't autosave! No main player found.");
+                Logging.LogError($"Couldn't autosave! No main player found.");
                 return;
             }
 
@@ -310,8 +418,8 @@ public class AutoSaver : BaseUnityPlugin
         }
         catch (Exception e)
         {
-            Logger.LogError($"Couldn't autosave!");
-            Logger.LogError($"Exception message: {e}");
+            Logging.LogError($"Couldn't autosave!");
+            Logging.LogError($"Exception message: {e}");
             return;
         }
     }