+++ AutoSaver/HarmonyPatches/Player_OnPlayerMapInstanceChange.cs
@@ -7,7 +7,7 @@ static class Player_OnPlayerMapInstanceChange
 {
     static void Postfix(Player __instance, MapInstance _new)
     {
-        if (__instance == Player._mainPlayer && (AutoSaver.Plugin.SaveOnMapChange || !AutoSaver.Plugin.CharacterActive))
+        if (__instance == Player._mainPlayer)
         {
             AutoSaver.Plugin.GameEntered();
         }