+++ AutoSaver/HarmonyReversePatches/CharacterAutoSaver.cs
@@ -75,11 +75,11 @@ static class CharacterAutoSaver
                 matcher.Advance(1);
             }
 
-            const int expectedLocations = 3;
+            const int expectedLocations = 2;
 
             if (patchLocations != expectedLocations)
             {
-                AutoSaver.Plugin.Logger.LogWarning($"WARNING: Expected {expectedLocations} patch locations, got {patchLocations}.");
+                AutoSaver.Plugin.Logger.LogWarning($"WARNING: CharacterAutoSaver expected {expectedLocations} patch locations, got {patchLocations}.");
                 AutoSaver.Plugin.Logger.LogWarning($"Either the vanilla code changed, or mods added extra stuff. This may or may not cause issues.");
             }